Spring-boot connection now ON.

+++ src/main/java/repository/ClientRepository.java
@@ -13,9 +13,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 
 @Repository
-public interface ClientRepository extends JpaRepository<Client, Integer> {
+public interface ClientRepository extends JpaRepository<Client, Long> {
 
     @Query("select c from Client c where c.name = :name")
     Client findByName(@Param("name") String name);
